Rename neural_trajectory to gpfa and refactor the codes accordingly
