+++ elephant/gpfa_src/README.md
@@ -3,3 +3,5 @@ Gaussian process factor analysis (GPFA), based on Byron Yu's MATLAB implementati
 
 The original MATLAB code is available at Byron Yu's website:
 https://users.ece.cmu.edu/~byronyu/software.shtml
+
+Extra requirements: scikit-leanr, tqdm
